Only check for scroll bouncing at top of page
the current logic is wrong; it should check the current scrollTop, not a cached one.
however, changing the logic to only check for bouncing at the top of the
page means we don’t need to start reading the document layout state in the
handleScroll function in order to fix the existing behavior.